+++ src/k8s_sandbox/_helm.py
@@ -158,6 +158,8 @@ async def uninstall(release_name: str, namespace: str, quiet: bool) -> None:
 
     The number of concurrent uninstall operations is limited by a semaphore.
 
+    "Release not found" errors are ignored.
+
     Args:
         release_name: The name of the Helm release to uninstall (e.g. abcdefgh).
         namespace: The Kubernetes namespace in which the release is installed.
@@ -184,11 +186,20 @@ async def uninstall(release_name: str, namespace: str, quiet: bool) -> None:
             if not quiet:
                 sys.stdout.write(result.stdout)
                 sys.stderr.write(result.stderr)
-            if not result.success:
+            # A helm uninstall failure with "release not found" implies that the release
+            # was never successfully installed or has already been uninstalled.
+            # When a helm release fails to install (or the user cancels the eval), this
+            # uninstall function will still be called, so these errors are common and
+            # result in error desensitisation.
+            if (
+                not result.success
+                and "release: not found" not in result.stderr.casefold()
+            ):
                 _raise_runtime_error(
                     "Helm uninstall failed.",
                     release=release_name,
                     namespace=namespace,
+                    returncode=result.returncode,
                     stdout=result.stdout,
                     stderr=result.stderr,
                 )